# docs/swagger.yaml
basePath: /v1
definitions:
...
  api.loginRequest:
    properties:
      email:
        type: string
      name:
        description: không bắt buộc, chỉ để phản hồi lại cho client
        type: string
      password:
        type: string
    type: object
  api.loginResponse:
    properties:
      email:
        type: string
      message:
        type: string
      name:
        type: string
    type: object
  api.registerRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  api.registerResponse:
    properties:
      email:
        type: string
      message:
        description: Token    string    `json:"token"`
        type: string
    type: object
...
  response.ErrorPayload:
    properties:
      code:
        type: string
      details:
        items:
          $ref: '#/definitions/response.ValidationErrorDetail'
        type: array
      message:
        type: string
      timestamp:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/response.ErrorPayload'
      success:
        type: boolean
    type: object
  response.Meta:
    properties:
      pagination:
        allOf:
        - $ref: '#/definitions/response.Pagination'
        description: Chỉ xuất hiện khi có phân trang, nếu không có sẽ tự động ẩn đi
      timestamp:
        type: string
    type: object
  response.Pagination:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  response.SuccessResponse:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
  response.ValidationErrorDetail:
    properties:
      field:
        type: string
      issue:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: Linh Community
    url: https://github.com/linhcommunity/nlscada-cloud
  description: Backend API for NL SCADA Cloud
  title: NL SCADA Center API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Xác thực người dùng và trả về JWT token
      parameters:
      - description: Thông tin đăng nhập
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Đăng nhập thành công
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/api.loginResponse'
              type: object
        "400":
          description: Lỗi dữ liệu
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Tài khoản hoặc mật khẩu không chính xác
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Đăng nhập
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Xóa cookie session_token để đăng xuất người dùng
      produces:
      - application/json
      responses:
        "200":
          description: Đăng xuất thành công
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "405":
          description: Phương thức không được hỗ trợ
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Đăng xuất
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Tạo token mới dựa trên refresh token (ở đây dùng cookie session_token)
      produces:
      - application/json
      responses:
        "200":
          description: Làm mới token thành công
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/api.loginResponse'
              type: object
        "401":
          description: Phiên đăng nhập không hợp lệ hoặc đã hết hạn
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Làm mới token
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Tạo người dùng mới. Không tạo site hay membership.
      parameters:
      - description: Thông tin đăng ký
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Đăng ký thành công
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/api.registerResponse'
              type: object
        "400":
          description: Lỗi dữ liệu
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Email đã tồn tại
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Đăng ký
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    description: Sử dụng "Bearer <token>", nhưng cookie session_token cũng được chấp
      nhận khi gọi từ browser
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"